Add command line interface to run the program
Moved the majority of program logic into a while loop
Added a end time argument to the deliver_packages function. This allows
us to take a snapshot and stop the algorithm at a particular point in
the simulation with a given time.

Provide a default value to the package class delivery status attribute